+++ lib/pronto/formatter/null_formatter.rb
@@ -1,0 +1,8 @@
+module Pronto
+  module Formatter
+    class NullFormatter
+      def format(_, _)
+      end
+    end
+  end
+end